refactor: realign extractor module boundaries
